+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   [refactor] Use prettier new configuration.
 -   [update] Se actualiza el archivo package.json
 -   [doc] Actualización de readme file.
+-   [feat] More security in helmet.
 
 ## v1.14.1